+++ src/Server/Process/Unix.php
@@ -12,10 +12,7 @@ use Innmind\TimeContinuum\{
     Period,
 };
 use Innmind\TimeWarp\Halt;
-use Innmind\Stream\{
-    Capabilities,
-    Watch,
-};
+use Innmind\Stream\Capabilities;
 
 /**
  * @internal
@@ -23,7 +20,6 @@ use Innmind\Stream\{
 final class Unix
 {
     private Clock $clock;
-    private Watch $watch;
     private Halt $halt;
     private Capabilities $capabilities;
     private Period $grace;
@@ -31,16 +27,14 @@ final class Unix
 
     public function __construct(
         Clock $clock,
-        Watch $watch,
-        Halt $halt,
         Capabilities $capabilities,
+        Halt $halt,
         Period $grace,
         Command $command,
     ) {
         $this->clock = $clock;
-        $this->watch = $watch;
-        $this->halt = $halt;
         $this->capabilities = $capabilities;
+        $this->halt = $halt;
         $this->grace = $grace;
         $this->command = $command;
     }
@@ -81,7 +75,6 @@ final class Unix
 
         return new Started(
             $this->clock,
-            $this->watch,
             $this->halt,
             $this->capabilities,
             $this->grace,